+++ src/Testing.Commons.NUnit/Constraints/ComposablePropertyConstraint.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq.Expressions;
 using NUnit.Framework.Constraints;
+using Testing.Commons.NUnit.Constraints.Support;
 
 namespace Testing.Commons.NUnit.Constraints;
