+++ src/Testing.Commons.NUnit/Constraints/NoPropertyChangedConstraint.Extensions.cs
