+++ src/Testing.Commons.NUnit/Constraints/NoPropertyChangedConstraint.cs
