+++ src/Testing.Commons.NUnit/Constraints/NoPropertyChangingConstraint.Extensions.cs
