+++ src/Testing.Commons.NUnit/Constraints/NoPropertyChangingConstraint.cs
