+++ src/Testing.Commons.NUnit/Constraints/PropertyChangedConstraint.Extensions.cs
