+++ src/Testing.Commons.NUnit/Constraints/PropertyChangedConstraint.cs
