+++ src/Testing.Commons.NUnit/Constraints/PropertyChangingConstraint.Extensions.cs
