+++ src/Testing.Commons.NUnit/Constraints/PropertyChangingConstraint.cs
