+++ src/Testing.Commons.NUnit/Constraints/RaisingConstraint.cs
@@ -2,6 +2,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq.Expressions;
 using NUnit.Framework;
 using NUnit.Framework.Constraints;
+using Testing.Commons.NUnit.Constraints.Support;
 
 namespace Testing.Commons.NUnit.Constraints;
