+++ src/Testing.Commons.NUnit/Constraints/SerializationConstraint.Extensions..cs
