+++ src/Testing.Commons.NUnit/Constraints/Support/ConstraintTesterBase.cs
@@ -10,12 +10,12 @@ namespace Testing.Commons.NUnit.Constraints.Support;
 public abstract class ConstraintTesterBase
 {
 	/// <summary>
-	/// Gets the error message of a failing contraint.
+	/// Gets the error message of a failing constraint.
 	/// </summary>
 	/// <param name="subject">Constraint to be tested.</param>
 	/// <param name="actual">The value to be tested.</param>
 	/// <typeparam name="T">Type of the tested value.</typeparam>
-	/// <returns>The message of a failing contraint or <see cref="string.Empty"/> if is not failing.</returns>
+	/// <returns>The message of a failing constraint or <see cref="string.Empty"/> if is not failing.</returns>
 	protected static string getMessage<T>([NotNull] Constraint subject, T actual)
 	{
 		string message = getMessage(subject.ApplyTo(actual));
@@ -23,12 +23,12 @@ public abstract class ConstraintTesterBase
 	}
 
 	/// <summary>
-	/// Gets the error message of a failing contraint.
+	/// Gets the error message of a failing constraint.
 	/// </summary>
 	/// <param name="subject">Constraint to be tested.</param>
 	/// <param name="actual">The action to be tested.</param>
 	/// <typeparam name="T">Type of the tested value.</typeparam>
-	/// <returns>The message of a failing contraint or <see cref="string.Empty"/> if is not failing.</returns>
+	/// <returns>The message of a failing constraint or <see cref="string.Empty"/> if is not failing.</returns>
 	protected static string getMessage<T>([NotNull] Constraint subject, ActualValueDelegate<T> actual)
 	{
 		string message = getMessage(subject.ApplyTo(actual));
