+++ src/Testing.Commons.NUnit/Constraints/Support/ExposingWriter.cs
