+++ tests/Testing.Commons.NUnit.Tests/Constraints/DateConstraintsTester.cs
@@ -1,5 +1,3 @@
-using System;
-using NUnit.Framework;
 using Testing.Commons.NUnit.Constraints;
 using Testing.Commons.Time;
